Break out login / signup js
